requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
cachetools==5.2.0
carla_birdeye_view==1.1.1
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.1
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.0.3
multidict==6.0.2
numpy==1.18.3
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pyvista==0.36.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.4.1
scooby==0.6.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
vtk==9.1.0
Werkzeug==2.2.2
wslink==1.6.6
yarl==1.8.1
zipp==3.10.0